+++ compat_tests/Cargo.toml
@@ -17,10 +17,10 @@ compat_0_8_1 = ["dep:openmls_0_8_1", "openmls/0-8-1-storage-format"]
 # Enabling `openmls_<version>/extensions-draft-08` adds the extensions-draft
 # storage-trait methods to that previous-version line's `StorageProvider`, which
 # our custom storage provider then implements (gated on `extensions-draft-compat`).
-extensions-draft = ["extensions-draft-current", "extensions-draft-compat"]
+extensions-draft = ["extensions-draft-compat", "extensions-draft-current"]
 extensions-draft-compat = [
-    "openmls_0_8_1?/extensions-draft-08",
     "openmls_0_7_4?/extensions-draft-08",
+    "openmls_0_8_1?/extensions-draft-08",
 ]
 extensions-draft-current = [
     "openmls/extensions-draft",
@@ -33,17 +33,19 @@ virtual-clients-draft = [
 
 storage_migration_0_7 = [
     "compat_0_7_4",
-    "openmls_traits_0_4_1",
-    "openmls/migration-test-utils",
-    "openmls_0_7_4?/migration-export",
     "dep:openmls_basic_credential_0_7",
+    "openmls/migration-import",
+    "openmls/test-utils",
+    "openmls_0_7_4?/migration-export",
+    "openmls_traits_0_4_1",
 ]
 storage_migration_0_8 = [
     "compat_0_8_1",
-    "openmls_traits_0_5_0",
-    "openmls/migration-test-utils",
-    "openmls_0_8_1?/migration-export",
     "dep:openmls_basic_credential_0_8",
+    "openmls/migration-import",
+    "openmls/test-utils",
+    "openmls_0_8_1?/migration-export",
+    "openmls_traits_0_5_0",
 ]
 
 [dependencies]
@@ -66,8 +68,8 @@ openmls_0_7_4 = { git = "https://github.com/openmls/openmls.git", branch = "open
 openmls_basic_credential_0_7 = { git = "https://github.com/openmls/openmls.git", branch = "openmls-v0.7.4-migration-helpers", package = "openmls_basic_credential", optional = true }
 openmls_traits_0_4_1 = { git = "https://github.com/openmls/openmls.git", branch = "openmls-v0.7.4-migration-helpers", package = "openmls_traits", optional = true }
 
-postcard = { version = "1.1.3", features = ["alloc"] }
 ciborium = "0.2"
+postcard = { version = "1.1.3", features = ["alloc"] }
 serde = { version = "^1.0" }
 serde_json = "1.0"
 thiserror = { version = "^2.0" }
