+++ openmls/src/group/public_group/mod.rs
@@ -68,7 +68,10 @@ mod validation;
 /// This struct holds all public values of an MLS group.
 #[derive(Debug)]
 #[cfg_attr(feature = "migration-import", derive(serde::Deserialize))]
-#[cfg_attr(feature = "migration-test-utils", derive(serde::Serialize))]
+#[cfg_attr(
+    all(feature = "test-utils", feature = "migration-import"),
+    derive(serde::Serialize)
+)]
 #[cfg_attr(any(test, feature = "test-utils"), derive(PartialEq, Clone))]
 pub struct PublicGroup {
     treesync: TreeSync,
